fix(landing): address review feedback (SSR split, auth store selectors, aria-pressed, CTA test)
- Extract CtaButton, ScreenshotsSection, EarlyAccessSection into separate "use client" files so the main LandingPage becomes a server component
- Use useAuthStore selectors in page.tsx and new client components to avoid full-store subscriptions
- Replace aria-current with aria-pressed on carousel dot buttons for correct semantics
- Add authenticated-user CTA test documenting that /trips/new is the destination when authenticated
